backend/requirements.txt
# Scraping et API Google PlayIA
google-play-scraper
google-generativeai

# Manipulation de données
pandas

# Base de données
sqlalchemy
python-dotenv
psycopg2-binary

# Tâches planifiées (Async)
celery
redis

# API
fastapi
uvicorn

# Auth
passlib[bcrypt]
python-jose[cryptography]
resend